\docType{data}
\name{gene1}
\alias{gene1}
\title{Gene Expression of Mice}
\format{A data frame with 76730 rows and 22 columns}
\usage{
data(gene1)
}
\description{
A dataset containing each gene as a row or variable with columns
corresponding to separate observations or measurements of that
gene across different samples. All of the data corresponds to
the level of expression for any particular gene. There is also
a final column indicating which regions of the brain that gene
is associated with. This is the format of a data frame
for input to the scorrplot package.
}
\source{Dr. Chris Gregg at the University of Utah}
\keyword{datasets}

