\docType{data}
\name{gene2}
\alias{gene2}
\title{Gene Expression of Mice, Part 2}
\format{A data frame with 53321 rows and 22 columns}
\usage{
data(gene2)
}
\description{
A dataset containing each gene as a row or variable with columns
corresponding to separate observations or measurements of that
gene across different samples. All of the data corresponds to
the level of expression for any particular gene. There is also
a final column indicating which regions of the brain that gene
is associated with, where this dataset differs from gene1 is
that there are up to nine different brain regions, as opposed
to only two. This is the format of a data frame
for input to the scorrplot package.
}
\source{Dr. Chris Gregg at the University of Utah}
\keyword{datasets}

