\docType{data}
\name{subway}
\alias{subway}
\title{Ridership on the NYC Subway System}
\format{A data frame with 423 rows and 103 columns}
\usage{
data(subway)
}
\description{
A dataset containing each subway stop as a row or variable
with the columns corresponding to each year. All the data are
counts of ridership at that stop for that year. There is also
a final column dictating which borough or region of NYC that
stop is associated with. This is the format of a data frame
for input to the scorrplot package.
}
\source{http://davis.wpi.edu/xmdv/datasets/subway.html}
\keyword{datasets}

